requirements.txt
faiss-cpu==1.10.*
google-cloud-logging==3.11.*
langchain==0.3.*
langchain-experimental==0.3.*
#langchain-google-vertexai>=2.0.14
# langchain-ollama>=0.2.3
pandas==2.*
pypdf==5.*
toml==0.10.*
tqdm==4.67.*
mypy==1.15.*
ipykernel==6.29.*
ragas==0.2.*
# langfuse==2.60.*
deepeval==2.7.*
chonkie[st]==1.4.*
openpyxl==3.1.*
python-dotenv==1.2.*
